RDKTV-17049 getSystemMemory to return values in Kb

+++ rdkshell.h
@@ -33,7 +33,7 @@ namespace RdkShell
     double seconds();
     double milliseconds();
     double microseconds();
-    bool systemRam(uint32_t &freeKb, uint32_t & totalKb, uint32_t& usedSwapKb);
+    bool systemRam(uint64_t &freeKb, uint64_t & totalKb, uint64_t& usedSwapKb);
     void setMemoryMonitor(const bool enable, const double interval);
     void setMemoryMonitor(std::map<std::string, RdkShellData> &configuration);
 }
